Refactor and rationalize
